[BROKEN]Work in progress attempt to move over to gravitation from arbitrary numbers
Currently generating the negative field does not work because it enters into a never ending while loop.

+++ Oliver-Chamberlain-Final-Project/Assets/Scripts/Agents/FlowFieldAgent.cs
@@ -15,8 +15,8 @@ public class FlowFieldAgent : MonoBehaviour {
     private Node recentlyVisitedNode;
 
 
-    public int bestcost;
-    public int currentNodeCost;
+    public double bestcost;
+    public double currentNodeCost;
 
 
     private void Start()
@@ -43,7 +43,7 @@ public class FlowFieldAgent : MonoBehaviour {
         Node agentNode = grid.NodeFromWorldPoint(transform.position);
         Node[] neighbourList = grid.GetNeighbours(agentNode).ToArray();
         Node bestNode = null;
-        int bestCost = 0;
+        double bestCost = 0;
          for(int i = 0; i < neighbourList.Length; i++)
         {
             if (!neighbourList[i].walkable || neighbourList[i] == null)
